+++ tests/merge/test_main.py
@@ -220,8 +220,8 @@ def test_additional_titles_badge(
     # expect title is visible and there are additional titles for 'r2'
     expect(resource_r2_result).to_contain_text(first_title.value)
     additional_title_badge = page.get_by_test_id("additional-titles-badge").first
-    additional_title_badge.scroll_into_view_if_needed()
     expect(additional_title_badge).to_be_visible()
+    additional_title_badge.scroll_into_view_if_needed()
     page.screenshot(path="tests_search_test_additional_titles_badge_on_visible.png")
     expect(additional_title_badge).to_have_text(
         build_ui_label_regex("components.titles.additional_titles")